Include -lgmp in the pkg-config file when building with gnutls
This is required when a user of librtmp is linking statically to
librtmp (when it is built with gnutls support).

Ideally everything of the Requires line should be in
Requires.private, and -lgmp would go into Libs.private - but if
only building a static lib, they should go where they are right
now instead. Keeping all of this where it is right now, for now,
for simplicity.