Set RTMP_PUB_ALLOC if tcUrl is allocated in PublisherAuth
This makes sure we don't free tcUrl in RTMP_Close when reconnecting,
and avoids leaks (or simplifies code to avoid leaks).

This fixes connecting with authentication (both adobe and limelight)
when the app has been specified separately from the rest of the url.

This is only a temporary step while refactoring, this is removed
in a later commit.